Refactor conversationFeed
