+++ client/src/app/components/messages/conversationFeed.tsx
@@ -1,16 +1,12 @@
-import React, { useRef, useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { Button, ProfilePic, Typography } from "..";
 import conversationFeedModule from "./css/conversationFeed.module.css";
 import { useDomPurify } from "@/app/hooks/useDomPurify";
-import { useRoomSocket } from "@/app/hooks/socket/useRoomSocket";
-import {
-  formatTime,
-  diffDateInSec,
-  displayRoomChatDate,
-} from "@/app/utils/date";
-import { SOCKETS_EMITTERS } from "@/socket/socketEvents";
+import { formatTime } from "@/app/utils/date";
+import { displayDate, displayUsernameAndProfilePic } from "./utils/date";
 import { useUser } from "@/app/controllers/userProvider";
-import { apiClient } from "@/app/utils/apiClient";
+import { useConversationUI } from "./utils/useMessageUI";
+import { useMessage } from "./utils/useMessage";
 import avatarPic from "@/assets/images/avatar.png"; // to delete once a default avatar pic is associated to a user if no profile pic.
 
 const ConversationFeed: React.FC<{
@@ -18,107 +14,39 @@ const ConversationFeed: React.FC<{
   contactProfilePic: string;
   handleInboxFeed: (roomId: string, content: string) => void;
 }> = ({ roomId, contactProfilePic, handleInboxFeed }) => {
-  const textRef = useRef<HTMLTextAreaElement>(null);
-  const bodyRef = useRef<HTMLDivElement>(null);
-  const messageRef = useRef<HTMLDivElement>(null);
-  const [textareaHeight, setTextareaHeight] = useState<number>(40);
-  const [bodyHeight, setBodyHeight] = useState<number>(250);
-  const { chatHistory, newMessage, isLoading, socket, sendMessageToServer } =
-    useRoomSocket(roomId);
-  const [messageQueue, setMessageQueue] = useState<any[] | []>(
-    chatHistory || []
-  );
-  const [messageContent, setMessageContent] = useState<string>("");
+  const {
+    textRef,
+    bodyRef,
+    textareaHeight,
+    bodyHeight,
+    adjustHeight,
+    scrollToBottom,
+  } = useConversationUI();
   const { user } = useUser();
+  const {
+    messageQueue,
+    messageContent,
+    isLoading,
+    chatHistory,
+    newMessage,
+    fetchMoreHistory,
+    registerMessage,
+    addMessage,
+  } = useMessage(roomId, user, bodyRef, textRef, handleInboxFeed);
+
   const { displayHTMLMessage } = useDomPurify();
-  let [offset, setOffset] = useState<number>(0);
-  const [allowFetchExtrahistory, setAllowFetchExtrahistory] =
-    useState<boolean>(false);
-  const allowFetchRef = useRef(allowFetchExtrahistory);
 
-  function adjustHeight() {
-    const textArea = textRef.current;
-    const body = bodyRef.current;
-    if (!textArea || !body) return;
+  useEffect(() => {
+    setTimeout(() => scrollToBottom(), 10);
+  }, [chatHistory]);
 
-    textArea.style.height = "auto";
-    const scrollHeight = textArea.scrollHeight;
-    const textareaMaxHeight = 100;
-
-    if (scrollHeight > textareaMaxHeight) {
-      setTextareaHeight(textareaMaxHeight);
-      setElementHeight(textArea, textareaMaxHeight, "auto");
-    } else {
-      setTextareaHeight(scrollHeight);
-      const newHeight = 250 - (scrollHeight - 40);
-      setElementHeight(body, newHeight);
-      setElementHeight(textArea, scrollHeight);
-      setBodyHeight(newHeight);
-    }
-  }
-  let controller: AbortController | null = null;
-  async function fetchMoreHistory() {
-    if (allowFetchRef.current === false) return;
-    const body = bodyRef.current;
-    if (!body) return;
-    const totalScrollHeight = body.scrollHeight;
-    const visibleHeight = body.clientHeight;
-    const amountScrolled = body.scrollTop;
-    const previousScrollHeight = body.scrollHeight;
-
-    const threshold = 0.01 * (totalScrollHeight - visibleHeight); // 80% scroll to the top
-    if (amountScrolled <= threshold) {
-      // User has scrolled up by 99.9%
-      setAllowFetchExtrahistory(false);
-      if (controller) {
-        controller.abort();
-      }
-      controller = new AbortController();
-      const signal: AbortSignal = controller.signal;
-      const olderChatHistory = await apiClient(
-        `message/${roomId}?offset=${offset + 1}`,
-        { signal }
-      );
-      setOffset(offset++);
-
-      if (olderChatHistory) {
-        setMessageQueue((prev) => [...olderChatHistory, ...prev]);
-
-        setTimeout(() => {
-          const newScrollHeight = body.scrollHeight;
-          const addedHeight = newScrollHeight - previousScrollHeight;
-          body.scrollTop = body.scrollTop + addedHeight;
-        }, 20);
-        setAllowFetchExtrahistory(true);
-
-        setTimeout(() => {}, 100);
-      } else {
-        setAllowFetchExtrahistory(false);
-      }
-    }
-  }
-
-  function scrollToBottomAtLoad() {
-    const body = bodyRef.current;
-    if (!body) return;
-    body.scrollTop = body.scrollHeight;
-  }
-
-  function scrollToBottom() {
-    const body = bodyRef.current;
-    if (body) {
-      const lastElement = body.lastChild;
-      if (lastElement) {
-        (lastElement as HTMLElement).scrollIntoView({ behavior: "smooth" });
-      }
-    }
-  }
+  useEffect(() => {
+    setTimeout(() => scrollToBottom(), 10);
+  }, [newMessage]);
 
   useEffect(() => {
     const textArea = textRef.current;
     const body = bodyRef.current;
-    scrollToBottomAtLoad();
-    setTimeout(() => setAllowFetchExtrahistory(true), 500);
     if (!textArea) return;
     const textarea = textRef.current;
     textarea.addEventListener("input", adjustHeight);
@@ -129,69 +57,11 @@ const ConversationFeed: React.FC<{
     return () => {
       textarea.removeEventListener("input", adjustHeight);
       body.removeEventListener("scroll", fetchMoreHistory);
-      socket.emit(SOCKETS_EMITTERS.LEAVE_ROOM, roomId);
     };
   }, []);
 
-  useEffect(() => {
-    if (newMessage) {
-      setMessageQueue((prev) => [
-        ...prev,
-        {
-          username: newMessage.username,
-          timestamp: newMessage.timestamp,
-          content: newMessage.message,
-          userId: newMessage.userId,
-        },
-      ]);
-      setTimeout(() => scrollToBottom(), 10);
-      handleInboxFeed(roomId, newMessage.message);
-    }
-  }, [newMessage]);
-
-  useEffect(() => {
-    if (chatHistory && offset === 0) {
-      setMessageQueue(chatHistory);
-      setTimeout(() => scrollToBottom(), 10);
-    } else if (chatHistory && offset > 0) {
-      const body = bodyRef.current;
-      setMessageQueue(() => [...chatHistory, ...messageQueue]);
-      if (!body) return;
-
-      setTimeout(() => {
-        messageRef.current?.scrollIntoView({ behavior: "smooth" });
-        setAllowFetchExtrahistory(true);
-      }, 500);
-    }
-  }, [chatHistory]);
-
-  useEffect(() => {
-    allowFetchRef.current = allowFetchExtrahistory;
-  }, [allowFetchExtrahistory]);
-
-  function registerMessage(e: React.ChangeEvent<HTMLTextAreaElement>) {
-    setMessageContent(e.target.value);
-  }
-
-  function addMessage() {
-    const body = bodyRef.current;
-    const textArea = textRef.current;
-    textArea!.value = "";
-    setElementHeight(textArea, 40);
-    setElementHeight(body, 250);
-    setMessageQueue((prev) => [
-      ...prev,
-      {
-        username: user!.username,
-        timestamp: new Date().toISOString(),
-        content: messageContent,
-        userId: user?.id,
-      },
-    ]);
-
-    sendMessageToServer(messageContent);
-    handleInboxFeed(roomId, messageContent);
-    setMessageContent("");
+  function handleSendMessage() {
+    addMessage();
     setTimeout(() => scrollToBottom(), 10);
   }
 
@@ -217,12 +87,9 @@ const ConversationFeed: React.FC<{
               NextMessage
             );
             return (
-              <>
+              <div key={idx}>
                 {time && <TimeStamp idx={idx} time={time} />}
-                <div
-                  className={conversationFeedModule.messageContainer}
-                  key={idx}
-                >
+                <div className={conversationFeedModule.messageContainer}>
                   {isDisplayed && (
                     <>
                       {" "}
@@ -246,7 +113,7 @@ const ConversationFeed: React.FC<{
                     paddingLeft: 60,
                   })}
                 </div>
-              </>
+              </div>
             );
           })
         )}
@@ -264,7 +131,7 @@ const ConversationFeed: React.FC<{
       <div className={conversationFeedModule.buttonContainer}>
         <Button
           disabled={messageContent.length > 0 ? false : true}
-          onClick={addMessage}
+          onClick={handleSendMessage}
           variant="primary"
           backgroundColor="black"
           padding={"5px 10px"}
@@ -299,53 +166,3 @@ const TimeStamp: React.FC<{ idx: number; time: string }> = ({ idx, time }) => {
     </div>
   );
 };
-
-function setElementHeight(
-  element: HTMLElement | null,
-  newHeight: number,
-  overflowY?: string
-) {
-  if (element) {
-    element.style.height = `${newHeight}px`;
-    if (overflowY) {
-      element.style.overflowY = overflowY;
-    }
-  }
-}
-
-function displayDate(currentMessageTime: string, prevMessageTime: any) {
-  if (!prevMessageTime) {
-    return currentMessageTime;
-  }
-  const dateToCompareFrom = prevMessageTime;
-
-  const response = displayRoomChatDate(currentMessageTime, dateToCompareFrom);
-
-  if (response.isToday) {
-    const diff = diffDateInSec(dateToCompareFrom, currentMessageTime);
-    // more than 15 min
-    if (diff > 60 * 15) {
-      return currentMessageTime;
-    }
-  } else if (response.isToday === false && response.continue === true) {
-    return currentMessageTime;
-  } else return null;
-}
-
-function displayUsernameAndProfilePic(
-  currentMessage: any,
-  prevMessage: any,
-  nextMessage: any
-) {
-  if (!prevMessage) return true;
-
-  if (
-    currentMessage.userId === prevMessage.userId ||
-    (!nextMessage && currentMessage.userId === prevMessage.userId)
-  ) {
-    const diff = diffDateInSec(prevMessage.timestamp, currentMessage.timestamp);
-    // 2 minutes or less
-    if (diff <= 60 * 2) return false;
-    else return true;
-  } else if (currentMessage.userId !== prevMessage.userId) return true;
-}
